requirements.txt
Pillow==8.1.2
Wand==0.6.6
